+++ Source/ACE.Entity/Enum/Properties/PropertyInstanceId.cs
@@ -1,6 +1,6 @@
 ﻿namespace ACE.Entity.Enum.Properties
 {
-    public enum PropertyInstanceId : uint
+    public enum PropertyInstanceId : ushort
     {
         // properties marked as ServerOnly are properties we never saw in PCAPs, from here:
         // http://ac.yotesfan.com/ace_object/not_used_enums.php
